* Fixed incorrect condition for ST_RIDING (follow-up r17110, bugreport:7254).
> Hercules merges:
* c69e4b6:
Added new item bonuses bFixedCast, bVariableCast & bFixedCastrate (see 'doc/item_bonus.txt' for info).
Added new conf for max_walk_path.

* bb61951:
Getting rid of unused indexes.
If the table has a multiple-column index, any leftmost prefix of the index can be used by the optimizer to find rows. For example, if you have a three-column index on (col1, col2, col3), you have indexed search capabilities on (col1), (col1, col2), and (col1, col2, col3).
Source: http://dev.mysql.com/doc/refman/5.5/en/multiple-column-indexes.html

* 7e4c578, 90eefa0:
Allow unsigned int values for guild exp (as it was before the change).
Atoll was introduced in a 2011 lib and vc9 and 10 don't have that, dropping to atol since we dont need the int64 and using a tricky conversion to get it to unsigned.

* a4802ea:
You no longer can dispell guild mates in non-pkable areas. (bugreport:6803)

* 54d0583:
Hells Plant no longer can be stacked on the ground. (bugreport:6825)

* 6fb7161:
Fire Pillar is now removed by land protector, and no longer can be cast over land protector. (bugreport:2177)

git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@17123 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/char/int_guild.c
@@ -601,9 +601,9 @@ static struct guild_castle* inter_guildcastle_fromsql(int castle_id)
 // Read exp_guild.txt
 static bool exp_guild_parse_row(char* split[], int column, int current)
 {
-	int exp = atoi(split[0]);
+	unsigned int exp = (unsigned int)atol(split[0]);
 
-	if (exp < 0 || exp >= INT_MAX) {
+	if (exp < 0 || exp >= UINT_MAX) {
 		ShowError("exp_guild: Invalid exp %d at line %d\n", exp, current);
 		return false;
 	}
